requirements.txt
openai
sentence-transformers
pypdf
pandas
python-dotenv
